libgdiplus: Fix build error relating to missing freetype/tttables.h, and update to libgdiplus 2.10.8
- libgdiplus 2.10 is fairly old, released 3 years ago (cea5a3f)
- libgdiplus 2.10.8 is curently the most recent release available on GitHub and is also old, released 2 years ago (5179dd3)
- there have been a substantial number of commits since these releases so it may be worth considering a recipe to build from a specific git commit (thoughts appreciated)

@see: https://github.com/mono/libgdiplus/releases

libgdiplus-2.10 and libgdiplus-2.10.8 recipes fail to build currently with an error

 gdiplus-private.h:33:31: fatal error: freetype/tttables.h: No such file or directory

This is a known issue addressed by commit 180c02e into master on 16/12/13. From the commit message,

"As of FreeType 2.1.6 (November 2003), using #include to include Freetype libraries directly is not supported.
This has come to a head, as in FreeType 2.5.0, the location of headers has been moved around, breaking building of libgdiplus.
This slight change uses the "official" way to include the required header file, without breaking building on older versions of the library."

Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>