mono: Update Mono to 3.2.8
- Mono 3.2.8 is a bug-fix only release based on 3.2.7 (19/02/2014)
- Mono 3.2.7 was a major forward release (17/01/2014)

Mono 3.2.7+ is particularly interesting because it has initial support
for ARM hardfp whereas until now we have had to build with vfp in order
to use Mono on ARM based boards.

Further release notes can be found here:

http://www.mono-project.com/Release_Notes_Mono_3.2

I have tested that Mono 3.2.8 builds and runs with a standard qemux86
target and with a vfp tuned i.MX6 target. hardfp remains to be tested.

I have also modified the maintainer list in the README to include myself.

Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>