mono-basic: Add VisualBasic.NET support with mono-basic v2.10
Synopsis:

@see http://www.mono-project.com/VisualBasic.NET_support for details on VB.Net support in Mono

Testing:

mono-basic recipe was added to a qemux86 image for testing. A simple VB helloworld application was used from this link

@see http://msdn.microsoft.com/en-us/library/3cf7t4xt(v=vs.110).aspx

The script was first compiled to an .exe on the build Ubuntu host, copied to the qemux86 image and executed successfully with

mono helloworld.exe

i.e. A message box was shown

Then the script was copied to the qemux86 image, compiled and executed successfully with

vbnc helloworld.vb
mono helloworld.exe

Note that when using VB.Net on an, e.g., Ubuntu host it may be necessary to add a missing link. The recipe handles creation of this link for the target package.

sudo ln -s /usr/lib/mono/4.0/Microsoft.VisualBasic.dll  /usr/lib/mono/4.5/Microsoft.VisualBasic.dll

@see http://mono.1490590.n4.nabble.com/mod-mono-with-VB-The-library-Microsoft-VisualBasic-dll-could-not-be-found-td4659980.html

Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>