mono-native: Further fixes to support build on host platform with host mono installed
Build failures were seen on an x64 host platform with Mono 3.2.1 installed. This seems to relate to the Mono build attempting to use the host installation of Mono, which this commit attempts to prevent with use of the EXTERNAL_MONO variable pointing to an invalid location

Testing: Now builds on Ubuntu 12.04 LTS x64 with Mono 3.2.1 installed in host environment. Also Ubuntu 12.04 x86

Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>