mono-4.xx.inc: disable parallel make
A race was observed during `make install` of mono-native under
PARALLEL_MAKE=-j6:

/usr/bin/install: cannot change permissions of
build/tmp-glibc/work/x86_64-linux/mono-native/4.0.2.4-r0/image/
build/tmp-glibc/sysroots/x86_64-linux/usr/lib/mono/xbuild-frameworks/
.NETFramework/v4.5/RedistList/FrameworkList.xml: No such file or
directory

This is definitely a race, because FrameworkList.xml was present once
the failed build completed.

The present mono build doesn't appear to build much of anything in
parallel in the first place, and I would guess that this issue applies
equally to the mono and mono-native builds. So the correct place for the
workaround ought to be in the shared .inc file.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>
Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>