mono-xsp: Add git recipe
The current release of mono-xsp (3.0.11) on github dates back to
May 2013. There are issues with this having a dependency on gmcs
which is no longer present in Mono.

There is a more recent release on the Mono download site (3.8)
dating to September 2014 but this has a similiar dependency on
gmcs.

The latest code in the github repo does address the gmcs
dependency.

Use this until a new release has been cut.

Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>