mono: Separate license from mono-4.xx.inc
During the last year Mono changed licenses from GPLv2 to MIT:
http://www.mono-project.com/news/2016/03/31/mono-relicensed-mit/

As such, we need to split the license out from mono-4.xx.inc so
that future consumers of that include can change the license as needed.

This changeset introduces mono-gplv2.inc to encode the GPLv2 license.
There is no functionality difference in this commit.

Signed-off-by: Barry Grussling <barry@grussling.com>