libgdiplus: remove libjpeg path from configure.ac
Building against Poky master, libgdiplus fails a Q/A check because
there is a -L/lib present in the do_compile step.

This was due to ${libjpeg_prefix} being dropped from ${LIBJPEG}
during configuration.

(There may be a way to prevent this being dropped but for now
removal of the search path setting works).

Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>