gtk-sharp: Add DIRFILES=1 to fix build issue under Yocto pyro
Build was failing with

Running transaction check
Transaction check succeeded.
Running transaction test
Error: Transaction check error:
   file /usr/lib/mono conflicts between attempted installs of gtk-sharp-2.12.21-r0.i586 and mono-libs-4.5-5.0.1.1-r0.i586
   file /usr/lib/mono/gac conflicts between attempted installs of gtk-sharp-2.12.21-r0.i586 and mono-gac-5.0.1.1-r0.i586

It appears this is related to an issue with the way the RPM
package manager deals with ownership of directories under pyro

From Zoltan Boszormenyi:

haven't Pyro replaced the RPM 5.x version with the more Fedora /
Red Hat / SuSE compatible 4.1x?

I vaguely remember that ownership of identical directories are not
allowed in two different RPM packages, it would be a problem on
Fedora, too.

RPM specfiles in Fedora can list empty directories with the %{dir}
directive (this is what different *filesystem* packages are for in
Fedora and RHEL/CentOS) but as far as I can see the Bitbake FILES_*
directives don't make a difference between file patterns and directories.

In Yocto, I only have experience with packaging IPKs. They list all
the parent directories of a file, too. E.g. for something like this:

FILES_${PN} += "${libdir}/package/somefile.bin"

the resulting IPK will contains these:

./usr
./usr/lib
./usr/lib/package
./usr/lib/package/somefile.bin

When you specify file patterns in an RPM specfile, the resulting RPM will
only contain the last one, i.e. the file with the full path. Only this
file will be owned by the RPM. Installing relies on the fact that the
whole path will be created anyway.

The installation conflict may come from RPM being stricter about directory
ownership and two packages can only list the same directories if both
the permissions and the creation date are the same. The latter would be
identical for two sub-packages of the same recipe but different for
different packages.

The problem seems to be inherent to how Bitbake creates the list of
packaged files.

...

The question is: how can we convince Bitbake to avoid adding
parent directories explicitly into the packages?

I think this is what you need to be aware of [...]

https://github.com/openembedded/openembedded-core/commit/0e33d232916125ba5305ced7200cc00f8b5f7b22

...

I suspect there is a deeper underlying issue as other recipes don't appear
to need to make use of DIRFILES=1. But this works as an interim solution

Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>