Add recipe for NuGet
NuGet is the package manager for dotNet / Mono and is used to
download already built 3rd party modules.

Two stable versions are provided: 3.5.0 and 4.1.0.
Preferred version is 3.5.0.

A nuget script is provided both for the package and
a cross-script to execute the native mono binary.

NuGet absolutely needs the SSL certificate store to be populated.
Either as root for the system-wide store under /usr/share/.mono:

$ sudo mozroots --import --machine --sync
$ sudo certmgr -ssl -m https://go.microsoft.com
$ sudo certmgr -ssl -m https://nugetgallery.blob.core.windows.net
$ sudo certmgr -ssl -m https://nuget.org

or for the Bitbake build user under $HOME/.config/.mono:

$ mozroots --import --sync
$ certmgr -ssl https://go.microsoft.com
$ certmgr -ssl https://nugetgallery.blob.core.windows.net
$ certmgr -ssl https://nuget.org

Since these commands don't populate the SSL certificate store
for BoringTLS introduced in Mono 4.8, either build Mono with
--disable-btls (like it's done for Mono 5.x in meta-mono) or
find out yourself how to populate the BTLS certificate store.

Signed-off-by: Zoltán Böszörményi <zboszor@pr.hu>