Only install files that are packaged
The do_install_append task copies files originating from the mono-native
package to the mono package. But since it copies the entire mono directory,
this causes problem if other mono related native packages are also installing
files into the same directory. In that case those files are not packaged, and
yocto complains with the QA Error "mono: Files/directories were installed but
not shipped in any package".

To workaround this problem, copy only those subdirectories that are actually
packaged.

Signed-off-by: Jef Driesen <jef.driesen@niko.eu>