Fix per-recipe sysroots in Yocto 2.3+
Yocto 2.3 (Pyro) added per-recipe sysroots. Previously a common
sysroot was used for every package installed. See:

https://www.yoctoproject.org/docs/2.3/ref-manual/ref-manual.html#moving-to-the-yocto-project-2.3-release

By default, ${libdir}/mono is not included in recipe-sysroot-native
for dependant packages which prevents running the native mono
binary that looks for DLLs relative to itself.

This patch should be backported to Pyro and Rocko from master.

Signed-off-by: Zoltán Böszörményi <zboszor@pr.hu>