mono: Workaround libmono-native.so packaging issue
Mono is expecting /usr/lib/libmono-native.so but this
was being packaged in mono-dev. When we package in the
mono-libs package we get a QA error as symlinks should
be in mono-dev.

I've not been able to change the build to require a
different library name, as this breaks mono-native
build so for now have packaged the .so symlink in
mono-libs and disabled the QA dev-so check.

Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>