mono: fix removing x86 binaries
We already have a fix to remove the objects but this fix depends on the
bash globbing mechanism. Since bitbake uses '#!/bin/sh' and not
'#!/bin/bash' this depends on your host. Ubuntu/Debian /bin/sh not
always refer to /bin/bash instead they are using /bin/dash which don't
support the globbing.

Replace the globbing with the mechanism used by commit ffa8c9992f8 which
is part of Mono 6.x. Also add a few more files to search for like the
commit did. While on it add message why this is needed and move it to
prepend so the configure phase uses clean dirs.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>