Fix mono 6.12.0.147 and add 6.12.0.154 (#95)
* mono: Rebase 6.12.0.147 to use the stable 6.12.0.122 tarball

mono 6.12.0.147 from git does not work properly, even if

  git submodule update --init --recursive

is run before do_configure. Something else is still missing
that would make mono to build and work from git.

Use the 6.12.0.122 stable tarball and cherry-pick most of the
patches from the 2020-02 (6.12.0.x) branch from git. The ones
that update git submodules' revisions are omitted.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>

* mono: Add recipe for version 6.12.0.154

It's done similarly to 6.12.0.147, so we start off of the
6.12.0.122 stable release tarball and add cherry-picked patches
from github.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>