mono: Add recipe for version 6.12.0.206
This is the last tagged version in https://github.com/mono/mono/tags
before the Mono was handed to the Wine project.
See also https://www.mono-project.com/ for the announcement.

The new version was not released as a tarball, so the bootstrap
procedure downloads the mini mono. This needs

  do_compile[network] = "1"

for the native build.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>